# docs/openapi.yaml
openapi: 3.0.3
info:
  title: nyxd control API
  description: |
    Minimal HTTP API for **nyxd**, served on a **Unix domain socket** (default `/run/nyxd/nyxd.sock`).
    There is no TCP listener by default.

    **Calling the API**

    ```bash
    curl -sS --unix-socket /run/nyxd/nyxd.sock http://localhost/v1/ping
    ```

    Replace `/run/nyxd/nyxd.sock` with your socket path or `NYXD_SOCKET` / `nyx -socket`.

    **Authentication:** none today (local trust). Socket permissions are typically `0660`.

    **Errors:** many routes use `text/plain` bodies from `http.Error`; JSON endpoints return `application/json` on success.
  version: "1.0.0"
  license:
    name: PolyForm Noncommercial License 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost
    description: |
      Use `curl --unix-socket <path> http://localhost/...` (host name is ignored; only the socket path matters).
...
paths:
...
  /v1/containers:
    get:
      tags: [containers]
      summary: List supervised containers
      description: |
        Without `detail`, returns `{"containers":["id",...]}` (IDs only).

        With `?detail=1` or `?detail=true`, also returns `items`: an array of
        `{id, image, ip, status}` suitable for `nyx ps`.
      operationId: listContainers
      parameters:
        - name: detail
          in: query
          required: false
          schema:
            type: string
            enum: ["1", "true", ""]
          description: When `1` or `true`, include per-container metadata in `items`.
      responses:
        "200":
          description: Container list (with or without detail)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ContainersResponse"
                  - $ref: "#/components/schemas/ContainersDetailResponse"
  /v1/images/pull:
    post:
      tags: [images]
      summary: Pull an image from a registry
      description: |
        When `stream` is **false** or omitted, returns a single JSON object on success.

        When `stream` is **true**, the response is **`application/x-ndjson`**: one JSON object per line
        (`PullEvent`), flushed progressively, ending with a `phase: done` line (or `phase: error`).
      operationId: pullImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullRequest"
      responses:
        "200":
          description: Success — either one JSON document or NDJSON stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullResponseJSON"
            application/x-ndjson:
              schema:
                type: string
                format: binary
                description: |
                  Each line is a `PullEvent` JSON object. Final line uses `phase: done` with summary fields.
        "400":
          description: Bad request (invalid body / missing ref)
        "502":
          description: Registry or transport error (non-stream mode)
        "503":
          description: Image store not configured
  /v1/containers/run:
    post:
      tags: [containers]
      summary: Create and start a container
      description: |
        Starts an OCI workload via crun. If the image is not in the local store, the daemon pulls it first.
        When `stream: true`, the response is `application/x-ndjson` (same `PullEvent` lines as `POST /v1/images/pull` if a pull runs, then `phase: run` with `container_id`). Otherwise the response is a single JSON `RunResponse`.
        If `id` is omitted, the server generates one from the image reference.
      operationId: runContainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Started (JSON) or pull+run stream (NDJSON when `stream` is true)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
            application/x-ndjson:
              schema:
                type: string
                format: binary
              description: |
                Same `PullEvent` lines as image pull when a registry fetch is required, including a final `phase: done` summary after pull, then a terminal `{"phase":"run","ok":true,"container_id":"...","ref":"..."}` line.
        "400":
          description: Validation or start error (plain text or NDJSON `phase:error` when streaming)
        "502":
          description: Registry or transport error during pull (plain text, or NDJSON when streaming)
        "503":
          description: Supervisor or image store unavailable
  /v1/containers/{id}/stop:
    post:
      tags: [containers]
      summary: Stop a container
      description: Sends the image stop signal (default SIGTERM) then SIGKILL if needed.
      operationId: stopContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stop accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopResponse"
        "400":
          description: Bad request
        "404":
          description: Container not found
        "503":
          description: Supervisor unavailable
  /v1/containers/{id}/remove:
    post:
      tags: [containers]
      summary: Remove a container
      description: |
        Stops the container if needed, deletes crun state, and tears down
        network and overlay resources.
      operationId: removeContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveResponse"
        "400":
          description: Bad request
        "404":
          description: Container not found
        "503":
          description: Supervisor unavailable
  /v1/containers/{id}/exec:
    post:
      tags: [containers]
      summary: Run a command inside a running container
      description: |
        Streams combined stdout/stderr as raw bytes (`application/octet-stream`).
        The HTTP connection stays open until the exec completes.

        Body variants:
        - `application/json` — `ExecRequest` only (stdin closed).
        - `application/x-nyxd-exec+v1` — first line is JSON `ExecRequest` plus newline; remaining bytes stream to `crun exec` stdin (used by `nyx exec -i`).
      operationId: execInContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecRequest"
          application/x-nyxd-exec+v1:
            schema:
              type: string
              format: binary
            description: JSON header line (ExecRequest) then raw stdin until EOF.
      responses:
        "200":
          description: Raw command output (stream)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request
        "503":
          description: Runtime unavailable
  /v1/containers/{id}/kill:
    post:
      tags: [containers]
      summary: Send a signal to a container (default SIGKILL)
      description: |
        Optional JSON body `{"signal":"KILL"}` (or `SIGKILL`). Used by foreground `nyx run` on Ctrl+C.
      operationId: killContainer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                signal:
                  type: string
                  description: Signal name without or with `SIG` prefix (e.g. `KILL`, `SIGTERM`).
      responses:
        "200":
          description: Kill accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KillResponse"
        "400":
          description: Bad request
        "404":
          description: Container not found
        "503":
          description: Supervisor unavailable
  /v1/containers/{id}/logs:
    get:
      tags: [containers]
      summary: Container stdout/stderr log file
      description: |
        Reads `logs/{id}.log` under the daemon data directory. With `plain=1` (default), JSONL
        lines are decoded to plain text. With `follow=1`, new lines are streamed until the client disconnects.
      operationId: containerLogs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tail
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 99999
          description: Number of trailing lines to send first (default 200).
        - name: follow
          in: query
          required: false
          schema:
            type: string
            enum: ["1", "true", ""]
        - name: plain
          in: query
          required: false
          schema:
            type: string
          description: When not `0`, decode JSONL log entries to text (default on).
      responses:
        "200":
          description: Plain text log stream (or snapshot when not following)
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: No log file for this container
        "500":
          description: Read error
  /v1/images:
    get:
      tags: [images]
      summary: List locally stored image references
      operationId: listImages
      responses:
        "200":
          description: Image ref list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagesListResponse"
        "503":
          description: Image store not configured
  /v1/images/remove:
    post:
      tags: [images]
      summary: Remove local image metadata for a ref
      operationId: removeImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageRemoveRequest"
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageRemoveResponse"
        "400":
          description: Bad request
        "503":
          description: Image store not configured
  /v1/images/prune:
    post:
      tags: [images]
      summary: Remove local image metadata not referenced by any supervised container
      description: |
        Uses in-memory supervisor container specs (`ContainerSpec.Image` strings) as the keep set.
        Does not delete content blobs under `blobs/sha256`.
      operationId: pruneImages
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagePruneRequest"
      responses:
        "200":
          description: Prune result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePruneResponse"
        "400":
          description: Bad request
        "503":
          description: Image store or supervisor not configured
components:
  schemas:
...
    ContainersResponse:
      type: object
      required: [containers]
      properties:
        containers:
          type: array
          items:
            type: string
    ContainersDetailResponse:
      type: object
      required: [containers, items]
      properties:
        containers:
          type: array
          items:
            type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContainerItem"
...
    RemoveResponse:
      type: object
      required: [ok, id]
      properties:
        ok:
          type: boolean
        id:
          type: string
    KillResponse:
      type: object
      required: [ok, id, signal]
      properties:
        ok:
          type: boolean
        id:
          type: string
        signal:
          type: string
          example: KILL
...
    PullRequest:
      type: object
      required: [ref]
      properties:
        ref:
          type: string
          description: Image reference (e.g. `nginx:alpine` or `registry/repo@sha256:...`)
          example: nginx:alpine
        stream:
          type: boolean
          description: If true, response is NDJSON stream instead of a single JSON object
    PullResponseJSON:
      type: object
      required: [ok, ref, config]
      properties:
        ok:
          type: boolean
        ref:
          type: string
        config:
          type: object
          description: Subset of OCI image config
          properties:
            os:
              type: string
            architecture:
              type: string
            entrypoint:
              type: array
              items:
                type: string
            cmd:
              type: array
              items:
                type: string
            working_dir:
              type: string
            env_len:
              type: integer
...
    RunResponse:
      type: object
      required: [ok, id, image]
      properties:
        ok:
          type: boolean
        id:
          type: string
        image:
          type: string
    StopResponse:
      type: object
      required: [ok, id]
      properties:
        ok:
          type: boolean
        id:
          type: string
    ExecRequest:
      type: object
      required: [argv]
      properties:
        argv:
          type: array
          items:
            type: string
          description: argv[0] is the program to exec